+++ crates/teloxide-core/src/types/allowed_update.rs
@@ -18,6 +18,7 @@ pub enum AllowedUpdate {
     CallbackQuery,
     ShippingQuery,
     PreCheckoutQuery,
+    PurchasedPaidMedia,
     Poll,
     PollAnswer,
     MyChatMember,