+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for TBA 7.7 ([#1357](https://github.com/teloxide/teloxide/pull/1357)) [**BC**]
 
+- Support for TBA 7.10 ([#1366](https://github.com/teloxide/teloxide/pull/1366))
+  - Add `filter_purchased_paid_media` filter
+
 ### Changed
 
 - New id types ([#1153](https://github.com/teloxide/teloxide/pull/1153)) [**BC**]